requirements.txt
flask
PyMySQL
python-dotenv
gunicorn